Better definition of the API, unit tests.
Beginning the long, arduous process of enumerating all the test cases I'd like to cover,
and continuing to document the API. One of the things that I also added was sanity checks
on the parameters so that requests fail early. There are also now unit tests in place to
cover interface failures, as well as the locking mechanism, and job completion.