Remove worker client reference.
The Qless::Worker does not need a reference to a client. It restrict
flexibility by coupling it to one client like this.  While it does not
normally make sense to do so, we have a case where would like a
single worker to pull jobs from queues on multiple qless servers,
which works just fine because the queue objects themselves know
what client to talk to.