Lua => LuaScript.
"LuaScript" is a better name for this class, as it represents a specific lua script, and not the lua language as a whole.

While I was at it, I moved the USING_LEGACY_REDIS_VERSION conditional out of the method and into the class body -- might as well evaulate that branch once at load time rather than once per script call.