Fix Subscriber to not use Thread#raise.
It's unsafe and "should not be used for any purpose":

http://headius.blogspot.com/2008/02/rubys-threadraise-threadkill-timeoutrb.html

In addition, this brings back the semantic the Subscriber
used to have, that it is guaranteed to be listening when
control is returned to the caller. This makes use of
the subscriber much easier to reason about.

Yes, `wait_until` is not particularly elegant.  We could probably
refactor to use a monitor condition variable.  For now this
brings back the semantics we want and removes the use of the
dangerous `Thread#raise`.