Remove Qless.worker_name.
With the changes in ef2d396055023eba66cc8b9b577abb73ec3ff1db,
it's no longer safe to use when using a forking worker,
because the worker name used by the client could differ
from what `Qless.worker_name` returns, given that it's
lazily initialized, so if it's called for the first
time in a child process, it'll have a different value
then what the client is using, causing problems.

Note: this isn't just a theoretic problem; it was happening
in plines, which referenced Qless.worker_name directly.