usage is different on OS X and Linux, apparently.
On my computer, it returns the memory usage in bytes.
On a linux server, I SSH'd into, it returns it in KB.

The OS X (BSD) getrusage man page says it is in bytes.
The Linux man page says it is in KB as of a particular
Linux version.

We deal with this by comparing the `ps` approach to
what rusage returns, and optionally dividing by 1024
if it appears that the amount is coming back in raw bytes.

We also switched everything to KB while we are at it
as that is easier to deal with.