Rename "failures" to "failed" in qless-stats
Graphite/whisper cannot have a name that serves both
as a metric name and as a dot-delimited prefix to
another metric. Because individual failure types are
reported as "failures.<type>", the "failures" metric
fails to be reported if ever there has been a report
of "failures.<type>". While this tool is not specific
to graphite, it is a common tool in the monitoring
ecosystem and it takes little effort to make this tool
compatible.